% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.hessSEM.assign.R
\name{ds.hessSEM.assign}
\alias{ds.hessSEM.assign}
\title{Combines values into a vector or list in the server-side}
\usage{
ds.hessSEM.assign(x = NULL, newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{a vector of character string providing the names of the objects to be combined.}

\item{newobj}{a character string that provides the name for the output object 
that is stored on the data servers. Default \code{c.newobj}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} 
objects obtained after login. If the \code{datasources} argument is not specified
the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
\code{ds.c} returns the vector of concatenating R
}
\description{
Concatenates objects into one vector.
}
\details{
To avoid combining the character names and not the
}
\author{
DataSHIELD Development Team
}
